Added vat support

+++ PosSystem/PosSystem/ProductClass.cs
@@ -14,9 +14,9 @@ namespace PosSystem
         public Int64 product_id = 0;
 
         // VAT = moms and is 12% for food and 25% for most other products
-        public int Vat = 0;
+        public Int64 Vat = 0;
 
-        public ProductClass(string name, decimal price, Int64 product_id ,string color, int vat = 12)
+        public ProductClass(string name, decimal price, Int64 product_id ,string color, Int64 vat = 12)
         {
             this.Name = name;
             this.Price = price;
